requirements.txt
asgiref==3.5.2
Django==4.1
django-extensions==3.2.0
djangorestframework==3.13.1
pytz==2022.2.1
sqlparse==0.4.2
asgiref==3.5.2
Django==4.1
django-crispy-forms==1.14.0
django-extensions==3.2.0
django-taggit==3.0.0
djangorestframework==3.13.1
pytz==2022.2.1
sqlparse==0.4.2
